+++ kite/src/main/kotlin/com/tang/kite/sql/LimitClause.kt
@@ -1,0 +1,12 @@
+package com.tang.kite.sql
+
+/**
+ * @author Tang
+ */
+class LimitClause(
+
+    val limit: Int,
+
+    val offset: Int
+
+)